add smoke tests, move tests to tests/e2e
