+++ .github/workflows/e2e-tests.yaml
@@ -25,7 +25,7 @@ jobs:
       max-parallel: 1
 
     env:
-      CONFIG_FILE: "tests/configs/${{ matrix.config-file }}"
+      CONFIG_FILE: "tests/e2e/configs/${{ matrix.config-file }}"
 
     steps:
       - name: Checkout
