+++ .github/workflows/pr-e2e-tests.yaml
@@ -22,7 +22,7 @@ jobs:
       max-parallel: 1
 
     env:
-      CONFIG_FILE: "tests/configs/${{ matrix.config-file }}"
+      CONFIG_FILE: "tests/e2e/configs/${{ matrix.config-file }}"
 
     steps:
       - name: Get PR branch
