+++ .github/workflows/pr-tests.yaml
@@ -17,7 +17,7 @@ jobs:
       max-parallel: 3
 
     env:
-      CONFIG_FILE: "tests/configs/${{ matrix.config-file }}"
+      CONFIG_FILE: "tests/e2e/configs/${{ matrix.config-file }}"
 
     steps:
       - name: Checkout
@@ -54,7 +54,7 @@ jobs:
             --cache_test_results=no \
             --test_output=all \
             --test_env=TEST_CONFIG_FILE=${{ env.CONFIG_FILE }} \
-            //tests:tests
+            //tests/e2e:tests
 
       # todo: change this to be post step of the action
       - name: Cleanup cluster
