+++ .github/workflows/smoke-tests.yaml
@@ -1,0 +1,115 @@
+name: Smoke tests
+
+on:
+  # Run every day at 18:00 UTC
+  schedule:
+    - cron: '0 18 * * *'
+  workflow_dispatch: {}
+  # debug ci for smoke tests
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  smoke-test-1:
+    strategy:
+      matrix:
+        config-file: [ "one-chain.yaml", "osmojs.yaml", "simapp.yaml", "multi-validator.yaml" ]
+      fail-fast: true
+      max-parallel: 4
+
+    env:
+      CONFIG_FILE: "tests/smoke/ci/${{ matrix.config-file }}"
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Setup Test infra
+        id: starship-action
+        uses: cosmology-tech/starship-action@0.2.14
+        with:
+          values: ${{ env.CONFIG_FILE }}
+          port-forward: true
+          version: 0.0.0
+          chart: ./charts/devnet
+
+      # todo: change this to be post step of the action
+      - name: Cleanup cluster
+        if: always()
+        run: |
+          helm delete $DEVNET_NAME --debug --namespace $DEVNET_NAMESPACE --wait || true
+          kubectl delete namespace $DEVNET_NAMESPACE --wait=true || true
+        env:
+          DEVNET_NAME: ${{ steps.starship-action.outputs.name }}
+          DEVNET_NAMESPACE: ${{ steps.starship-action.outputs.namespace }}
+
+  smoke-test-2:
+    needs: [smoke-test-1]
+
+    strategy:
+      matrix:
+        config-file: [ "one-chain.yaml", "osmojs.yaml", "simapp.yaml", "multi-validator.yaml" ]
+      fail-fast: true
+      max-parallel: 4
+
+    env:
+      CONFIG_FILE: "tests/smoke/ci/${{ matrix.config-file }}"
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Setup Test infra
+        id: starship-action
+        uses: cosmology-tech/starship-action@0.2.14
+        with:
+          values: ${{ env.CONFIG_FILE }}
+          port-forward: true
+          version: 0.0.0
+          chart: ./charts/devnet
+
+      # todo: change this to be post step of the action
+      - name: Cleanup cluster
+        if: always()
+        run: |
+          helm delete $DEVNET_NAME --debug --namespace $DEVNET_NAMESPACE --wait || true
+          kubectl delete namespace $DEVNET_NAMESPACE --wait=true || true
+        env:
+          DEVNET_NAME: ${{ steps.starship-action.outputs.name }}
+          DEVNET_NAMESPACE: ${{ steps.starship-action.outputs.namespace }}
+
+  smoke-test-3:
+    needs: [smoke-test-2]
+
+    strategy:
+      matrix:
+        config-file: [ "one-chain.yaml", "osmojs.yaml", "simapp.yaml", "multi-validator.yaml" ]
+      fail-fast: true
+      max-parallel: 4
+
+    env:
+      CONFIG_FILE: "tests/smoke/ci/${{ matrix.config-file }}"
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Setup Test infra
+        id: starship-action
+        uses: cosmology-tech/starship-action@0.2.14
+        with:
+          values: ${{ env.CONFIG_FILE }}
+          port-forward: true
+          version: 0.0.0
+          chart: ./charts/devnet
+
+      # todo: change this to be post step of the action
+      - name: Cleanup cluster
+        if: always()
+        run: |
+          helm delete $DEVNET_NAME --debug --namespace $DEVNET_NAMESPACE --wait || true
+          kubectl delete namespace $DEVNET_NAMESPACE --wait=true || true
+        env:
+          DEVNET_NAME: ${{ steps.starship-action.outputs.name }}
+          DEVNET_NAMESPACE: ${{ steps.starship-action.outputs.namespace }}
