+++ tests/e2e/e2e_test.go
@@ -1,4 +1,4 @@
-package tests
+package e2e
 
 import (
 	"fmt"
@@ -8,12 +8,10 @@ import (
 	"strings"
 	"testing"
 
+	pb "github.com/cosmology-tech/starship/exposer/exposer"
 	"github.com/golang/protobuf/jsonpb"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
-
-	pb "github.com/cosmology-tech/starship/exposer/exposer"
 )
 
 var configEnvKey = "TEST_CONFIG_FILE"
