+++ tests/e2e/registry_test.go
@@ -1,4 +1,4 @@
-package tests
+package e2e
 
 import (
 	"fmt"
