+++ tests/smoke/ci/one-chain.yaml
@@ -1,0 +1,27 @@
+chains:
+  - name: osmosis-1
+    type: osmosis
+    numValidators: 1
+    ports:
+      rest: 1313
+      rpc: 26653
+      exposer: 38083
+    resources:
+      cpu: "0.5"
+      memory: 500M
+    faucet:
+      concurrency: 2
+
+registry:
+  enabled: true
+  ports:
+    rest: 8081
+    grpc: 9091
+  resources:
+    cpu: "0.1"
+    memory: 100M
+
+resources:
+  wait:
+    cpu: "0.1"
+    memory: 100M
