+++ tests/smoke/ci/osmojs.yaml
@@ -1,0 +1,52 @@
+chains:
+  - name: osmosis-1
+    type: osmosis
+    image: pyramation/osmosis:v16.1.0
+    numValidators: 1
+    ports:
+      rest: 1317
+      rpc: 26657
+      faucet: 8007
+    resources:
+      cpu: "0.2"
+      memory: "200M"
+  - name: cosmos-2
+    type: cosmos
+    numValidators: 1
+    ports:
+      rest: 1313
+      rpc: 26653
+      faucet: 8003
+    resources:
+      cpu: "0.2"
+      memory: "200M"
+
+relayers:
+  - name: osmos-cosmos
+    type: hermes
+    replicas: 1
+    chains:
+      - osmosis-1
+      - cosmos-2
+    resources:
+      cpu: "0.1"
+      memory: "100M"
+
+registry:
+  enabled: true
+  ports:
+    rest: 8081
+    grpc: 9091
+  resources:
+    cpu: "0.1"
+    memory: "100M"
+
+exposer:
+  resources:
+    cpu: "0.1"
+    memory: "100M"
+
+faucet:
+  resources:
+    cpu: "0.1"
+    memory: "100M"
