+++ tests/smoke/ci/simapp.yaml
@@ -1,0 +1,35 @@
+chains:
+  - name: simapp
+    type: simapp
+    image: ghcr.io/cosmos/simapp:latest
+    numValidators: 2
+    ports:
+      rest: 1317
+      rpc: 26657
+      grpc: 9091
+      faucet: 8001
+    resources:
+      cpu: "0.2"
+      memory: "400M"
+    faucet:
+      # use v0.31.0-alpha.2 cosmjs-faucet for compatibility with simapp v0.47+
+      # todo: faucet seems to throw error with sdk version 0.50, still seems to work
+      image: ghcr.io/cosmology-tech/starship/cosmjs-faucet:v0.31.0
+      concurrency: 2
+      resources:
+        cpu: "0.1"
+        memory: "200M"
+
+registry:
+  enabled: true
+  ports:
+    rest: 8081
+  resources:
+    cpu: "0.1"
+    memory: "100M"
+
+exposer:
+  image: ghcr.io/cosmology-tech/starship/exposer:20230808-2333929
+  resources:
+    cpu: "0.1"
+    memory: "100M"
